Fixed #165, do not register Firework entities.
By doing so, all other Vanilla entities will be able to be saved as chunks. Fireworks entities in this plugin are just temporary and will be destroyed after the server has been stopped or match has been ended. However this feature is yet to be added soon in the future.